deploy/cloud_function/requirements.txt
functions-framework>=3.5
google-auth>=2.31
requests>=2.32
